Refactored landing box into separate component.
